Set cookie only if cookie.value exists
Added an additional IF statement that checks if the cookie.value exists. Currently, an empty cookie is set by the tag in the response headers even if the cookie.value is undefined. Maybe I am missing some context here but it looks more convenient to have the tag handle this and skip cookies that would be empty (rather than try to avoid this with triggers)